+++ src/components/layer-controller/VectorLayerController.js
@@ -1,0 +1,58 @@
+import React, { useCallback, useState, useReducer } from 'react';
+
+import Grid from '@material-ui/core/Grid';
+import Checkbox from '@material-ui/core/Checkbox';
+import Paper from '@material-ui/core/Paper';
+import Typography from '@material-ui/core/Typography';
+import Slider from '@material-ui/core/Slider';
+
+import debounce from 'lodash/debounce';
+
+import { useExpansionPanelStyles } from './styles';
+
+export default function VectorLayerController(props) {
+  const [slider, setSliderValue] = useState(1);
+  const [isOn, toggle] = useReducer(v => !v, true);
+
+  const { label, handleOpacityChange, handleToggleChange } = props;
+
+  const handleSliderChange = v => isOn && (handleOpacityChange(v) && setSliderValue(v));
+  const handleCheckBoxChange = v => handleToggleChange(v) && toggle();
+  const handleSliderChangeDebounced = useCallback(
+    debounce(handleSliderChange, 3, { trailing: true }),
+    [isOn],
+  );
+
+  const classes = useExpansionPanelStyles();
+  return (
+    <Grid item style={{ marginTop: '10px' }}>
+      <Paper className={classes.root}>
+        <Typography
+          style={{
+            paddingTop: '15px',
+            paddingLeft: '10px',
+            marginBottom: '-5px',
+          }}
+        >
+          {label}
+        </Typography>
+        <Grid container direction="row" justify="space-between">
+          <Grid item xs={2}>
+            <Checkbox color="primary" checked={isOn} onChange={(e, v) => handleCheckBoxChange(v)} />
+          </Grid>
+          <Grid item xs={9} style={{ paddingRight: '8px' }}>
+            <Slider
+              value={slider}
+              min={0}
+              max={1}
+              step={0.001}
+              onChange={(e, v) => handleSliderChangeDebounced(v)}
+              style={{ marginTop: '7px' }}
+              orientation="horizontal"
+            />
+          </Grid>
+        </Grid>
+      </Paper>
+    </Grid>
+  );
+}