+++ src/components/scatterplot/Scatterplot.js
@@ -1,5 +1,8 @@
-import React, { useState, useRef, useCallback } from 'react';
+import React, {
+  useState, useRef, useCallback, useMemo,
+} from 'react';
 import DeckGL, { OrthographicView } from 'deck.gl';
+import { quadtree } from 'd3-quadtree';
 import { SelectableScatterplotLayer, getSelectionLayers } from '../../layers';
 import ToolMenu from '../ToolMenu';
 import {
@@ -62,7 +65,9 @@ export default function Scatterplot(props) {
         throw new Error(`Expected to find "${mapping}", but available mappings are: ${available}`);
       }
       const mappedCell = mappings[mapping];
-      return [mappedCell[0], mappedCell[1], 0];
+      // The negative applied to the y-axis is because
+      // graphics rendering has the y-axis positive going south.
+      return [mappedCell[0], -mappedCell[1], 0];
     },
     getCellColor = cellEntry => (cellColors && cellColors[cellEntry[0]]) || DEFAULT_COLOR,
     getCellIsSelected = cellEntry => (
@@ -118,32 +123,65 @@ export default function Scatterplot(props) {
     updateViewInfo(viewRef.current);
   }, [viewRef, updateViewInfo]);
 
-  if (cells) {
-    clearPleaseWait('cells');
-  }
+  const cellsData = useMemo(() => {
+    let result = null;
+    if (cells) {
+      // Process cells data and cache into re-usable array.
+      result = Object.entries(cells);
+      if (clearPleaseWait) clearPleaseWait('cells');
+    }
+    return result;
+  }, [cells, clearPleaseWait]);
 
-  const layers = (cells ? [
-    new SelectableScatterplotLayer({
-      id: CELLS_LAYER_ID,
-      backgroundColor: (theme === 'dark' ? [0, 0, 0] : [241, 241, 241]),
-      isSelected: getCellIsSelected,
-      opacity: cellOpacity,
-      radiusScale: cellRadiusScale,
-      radiusMinPixels: 1.5,
-      radiusMaxPixels: 10,
-      getPosition: getCellPosition,
-      getColor: getCellColor,
-      onClick: (info) => {
-        if (tool) {
-          // If using a tool, prevent individual cell selection.
-          // Let SelectionLayer handle the clicks instead.
-          return;
-        }
-        onCellClick(info);
-      },
-      ...cellLayerDefaultProps(Object.entries(cells), updateStatus, updateCellsHover, uuid),
-    }),
-  ] : []);
+  // Graphics rendering has the y-axis positive going south,
+  // so we need to flip it for rendering tooltips.
+  const flipYTooltip = true;
+
+  const layers = useMemo(() => {
+    if (!cellsData) {
+      return [];
+    }
+    return [
+      new SelectableScatterplotLayer({
+        id: CELLS_LAYER_ID,
+        backgroundColor: (theme === 'dark' ? [0, 0, 0] : [241, 241, 241]),
+        isSelected: getCellIsSelected,
+        opacity: cellOpacity,
+        radiusScale: cellRadiusScale,
+        radiusMinPixels: 1.5,
+        radiusMaxPixels: 10,
+        getPosition: getCellPosition,
+        getColor: getCellColor,
+        onClick: (info) => {
+          if (tool) {
+            // If using a tool, prevent individual cell selection.
+            // Let SelectionLayer handle the clicks instead.
+            return;
+          }
+          onCellClick(info);
+        },
+        ...cellLayerDefaultProps(
+          cellsData, updateStatus, updateCellsHover, uuid, flipYTooltip,
+        ),
+      }),
+    ];
+  }, [cellsData, theme, getCellIsSelected, cellOpacity, cellRadiusScale,
+    getCellPosition, getCellColor, updateStatus, updateCellsHover, uuid,
+    tool, onCellClick, flipYTooltip]);
+
+  const cellsQuadTree = useMemo(() => {
+    // Use the cellsData variable since it is already
+    // an array, converted by Object.entries().
+    if (!cellsData) {
+      // Abort if the cells data is not yet available.
+      return null;
+    }
+    const tree = quadtree()
+      .x(d => getCellCoords(d[1])[0])
+      .y(d => getCellCoords(d[1])[1])
+      .addAll(cellsData);
+    return tree;
+  }, [getCellCoords, cellsData]);
 
   const selectionLayers = getSelectionLayers(
     tool,
@@ -151,6 +189,8 @@ export default function Scatterplot(props) {
     CELLS_LAYER_ID,
     getCellCoords,
     updateCellsSelection,
+    cellsQuadTree,
+    flipYTooltip,
   );
 
   const deckProps = {