+++ src/components/utils.js
@@ -1,4 +1,5 @@
 import { useRef, useState, useEffect } from 'react';
+// eslint-disable-next-line vitessce-rules/prevent-pubsub-import
 import PubSub from 'pubsub-js';
 import debounce from 'lodash/debounce';
 import { COORDINATE_SYSTEM } from 'deck.gl';
@@ -11,7 +12,7 @@ export function makeCellStatusMessage(cellInfoFactors) {
   ).join('; ');
 }
 
-export function cellLayerDefaultProps(cells, updateStatus, updateCellsHover, uuid) {
+export function cellLayerDefaultProps(cells, updateStatus, updateCellsHover, uuid, flipY = false) {
   return {
     coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
     data: cells,
@@ -25,10 +26,17 @@ export function cellLayerDefaultProps(cells, updateStatus, updateCellsHover, uui
       if (info.object) {
         const [cellId, cellInfo] = info.object;
         const { factors = {}, xy, mappings = {} } = cellInfo;
+        const scatterplotMappings = { ...mappings };
+        if (flipY) {
+          Object.keys(scatterplotMappings).forEach((mapping) => {
+            const arr = [scatterplotMappings[mapping][0], -scatterplotMappings[mapping][1]];
+            scatterplotMappings[mapping] = arr;
+          });
+        }
         updateStatus(makeCellStatusMessage(factors));
         updateCellsHover({
           cellId,
-          mappings: { xy, ...mappings },
+          mappings: { xy, ...scatterplotMappings },
           uuid,
           factors,
         });