+++ src/events.js
@@ -1,7 +1,10 @@
 const ADD = '.add';
 const CHANGE = '.change';
 const REMOVE = '.remove';
+const OPACITY = '.opacity';
+const ON = '.on';
 
+export const RESET = 'reset';
 export const CLEAR_PLEASE_WAIT = 'clear_please_wait';
 
 export const VIEW_INFO = 'view.info';
@@ -13,6 +16,8 @@ export const CELLS_ADD = CELLS + ADD;
 export const CELLS_SELECTION = `${CELLS}.selection`;
 export const CELLS_COLOR = `${CELLS}.color`;
 export const CELLS_HOVER = `${CELLS}.hover`;
+export const CELLS_SET_OPACITY = CELLS + OPACITY;
+export const CELLS_TURN_ON = CELLS + ON;
 
 export const CELL_SETS = `${CELLS}.sets`;
 export const CELL_SETS_ADD = CELL_SETS + ADD;
@@ -31,6 +36,8 @@ export const GENES_ADD = GENES + ADD;
 
 export const MOLECULES = 'molecules';
 export const MOLECULES_ADD = MOLECULES + ADD;
+export const MOLECULES_SET_OPACITY = MOLECULES + OPACITY;
+export const MOLECULES_TURN_ON = MOLECULES + ON;
 
 export const NEIGHBORHOODS = 'neighborhoods';
 export const NEIGHBORHOODS_ADD = NEIGHBORHOODS + ADD;